+++ src/extensions.ts
@@ -1,7 +1,6 @@
 import * as path from 'path';
 import * as cf from '@aws-cdk/aws-cloudfront';
 import * as lambda from '@aws-cdk/aws-lambda';
-import * as iam from '@aws-cdk/aws-iam';
 import { NodejsFunction } from '@aws-cdk/aws-lambda-nodejs';
 import * as cdk from '@aws-cdk/core';
 import { ServerlessApp } from './';
@@ -246,11 +245,13 @@ function bumpFunctionVersion(scope: cdk.Construct, id: string, functionArn: stri
 export class DefaultDirIndex extends Custom {
   readonly lambdaFunction: lambda.Version;
   constructor(scope: cdk.Construct, id: string) {
-
-    super(scope, id, {
+    const func = new NodejsFunction(scope, 'DefaultDirIndexFunc', {
+      entry: `${EXTENSION_ASSETS_PATH}/cf-default-dir-index/index.ts`,
+      // L@E does not support NODE14 so use NODE12 instead.
       runtime: lambda.Runtime.NODEJS_12_X,
-      handler: 'index.handler',
-      code: lambda.AssetCode.fromAsset(`${EXTENSION_ASSETS_PATH}/cf-default-dir-index`),
+    });
+    super(scope, id, {
+      func,
       eventType: cf.LambdaEdgeEventType.ORIGIN_REQUEST,
       solutionId: 'SO8134',
       templateDescription: 'Cloudfront extension with AWS CDK - Default Directory Index for Amazon S3 Origin.',
@@ -258,6 +259,88 @@ export class DefaultDirIndex extends Custom {
     this.lambdaFunction = this.functionVersion;
   }
 };
+
+/**
+ * Display customized error pages, or mask 4XX error pages, based on where the error originated
+ *
+ *  use case - see https://aws.amazon.com/blogs/networking-and-content-delivery/customize-403-error-pages-from-amazon-cloudfront-origin-with-lambdaedge/
+ */
+export class RedirectCustomErrorPage extends Custom {
+  readonly lambdaFunction: lambda.Version;
+  constructor(scope: cdk.Construct, id: string) {
+
+    super(scope, id, {
+      runtime: lambda.Runtime.PYTHON_3_7,
+      handler: 'index.handler',
+      code: lambda.AssetCode.fromAsset(`${EXTENSION_ASSETS_PATH}/redirect-custom-error-page`),
+      eventType: cf.LambdaEdgeEventType.ORIGIN_RESPONSE,
+      solutionId: '',
+      templateDescription: 'Cloudfront extension with AWS CDK - Display customized error pages, or mask 4XX error pages, based on where the error originated.',
+    });
+    this.lambdaFunction = this.functionVersion;
+  }
+};
+
+export interface SelectOriginByViwerCountryProps {
+  /**
+   * The pre-defined country code table.
+   * Exampe: { 'US': 'amazon.com' }
+   */
+  readonly countryTable: { [code: string]: string };
+}
+
+/**
+ * selective origin by viewer counry
+ */
+export class SelectOriginByViwerCountry extends Custom {
+  constructor(scope: cdk.Construct, id: string, props: SelectOriginByViwerCountryProps) {
+    const func = new NodejsFunction(scope, 'SelectOriginViewerCountryFunc', {
+      entry: `${EXTENSION_ASSETS_PATH}/select-origin-by-viewer-country/index.ts`,
+      // L@E does not support NODE14 so use NODE12 instead.
+      runtime: lambda.Runtime.NODEJS_12_X,
+      bundling: {
+        define: {
+          'process.env.COUNTRY_CODE_TABLE': jsonStringifiedBundlingDefinition(props.countryTable),
+        },
+      },
+    });
+    super(scope, id, {
+      func,
+      eventType: cf.LambdaEdgeEventType.ORIGIN_REQUEST,
+      solutionId: '',
+      templateDescription: 'Cloudfront extension with AWS CDK - Selective Origin by Viewer Country',
+    });
+  }
+};
+
+export interface AccessOriginByGeolocationProps {
+  /**
+   * The pre-defined country code table.
+   * Exampe: { 'US': 'amazon.com' }
+   */
+   readonly countryTable: { [code: string]: string };
+}
+export class AccessOriginByGeolocation extends Custom {
+  readonly lambdaFunction: lambda.Version;
+  constructor(scope: cdk.Construct, id: string, props: AccessOriginByGeolocationProps) {
+    const func = new NodejsFunction(scope, 'CustomFunc', {
+      entry: `${EXTENSION_ASSETS_PATH}/cf-access-origin-by-geolocation/index.ts`,
+      handler: 'handler',
+      bundling: {
+        define: {
+          'process.env.COUNTRY_CODE_TABLE': jsonStringifiedBundlingDefinition(props.countryTable)
+        }
+      },
+    })
+    super(scope, id, {
+      func: func,
+      eventType: cf.LambdaEdgeEventType.ORIGIN_REQUEST,
+      solutionId: '',
+      templateDescription: 'Cloudfront extension with AWS CDK - access Origin by geolocation',
+    });
+    this.lambdaFunction = this.functionVersion;
+  }
+}
 
 /**
  * Simple content generation
@@ -279,43 +362,8 @@ export class SimpleLambdaEdge extends Custom {
   }
 };
 
-export interface AccessOriginByGeolocationProps {
-  readonly bucketDomainName: string;
-}
-export class AccessOriginByGeolocation extends Custom {
-  readonly lambdaFunction: lambda.Version;
-  constructor(scope: cdk.Construct, id: string, props: AccessOriginByGeolocationProps) {
-    const func = new NodejsFunction(scope, 'CustomFunc', {
-      entry: path.resolve(__dirname, '..', 'lambda-assets', 'extensions', 'cf-access-origin-by-geolocation/index.js'),
-      handler: 'handler',
-      bundling: {
-        define: {
-          'MY_ENV': JSON.stringify({a: 'x', b: 'y', c: 'z'})
-            .replace(/"/g, '\\"')
-            .replace(/,/g, '\\,')
-        }
-      },
-    })
-    console.log(props)
-    super(scope, id, {
-      func: func,
-      eventType: cf.LambdaEdgeEventType.ORIGIN_REQUEST,
-    });
-    this.lambdaFunction = this.functionVersion;
-    const customConfigFunctionRole = new iam.Role(scope, 'CustomConfigFunctionRole', {
-      assumedBy: new iam.CompositePrincipal(
-        new iam.ServicePrincipal('lambda.amazonaws.com'),
-      )
-    });
-    customConfigFunctionRole.addManagedPolicy(iam.ManagedPolicy.fromAwsManagedPolicyName('AWSLambda_FullAccess'))
-    new NodejsFunction(scope, 'CustomConfigFunction', {
-      entry: path.resolve(__dirname, '..', 'lambda-assets', 'custom', 'cf-access-origin-by-geolocation/index.js'),
-      handler: 'handler',
-      role: customConfigFunctionRole,
-      environment: {
-        EXTEND_FUNCTION_NAME: func.functionName,
-        EXTEND_FUNCTION_VERSION: this.lambdaFunction.version
-      }
-    })
-  }
+function jsonStringifiedBundlingDefinition(value: any): string {
+  return JSON.stringify(value)
+    .replace(/"/g, '\\"')
+    .replace(/,/g, '\\,');
 }